% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_model_par.R
\name{average_over}
\alias{average_over}
\title{Average across factors in a data.frame}
\usage{
average_over(df, varname, collapse_factors, weight = NULL, auto_weight = TRUE)
}
\arguments{
\item{df}{a \code{data.frame}}

\item{varname}{the variable to be averaged}

\item{collapse_factors}{the variables to average across}

\item{weight}{}
}
\value{
\code{data.frame} in which the variable named by \code{varname} was averaged
across levels of
}
\description{
Average across factors in a data.frame
}
