% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blms-package.R
\docType{package}
\name{blms}
\alias{blms-package}
\alias{blms}
\title{blms}
\description{
Fit Bayesian learning models using Stan.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simon-desch/blms}
  \item Report bugs at \url{https://github.com/simon-desch/blms/issues}
}

}
\author{
\strong{Maintainer}: Simon Desch \email{simon.desch@gmx.de}

}
