% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blms_model.R
\name{blms_model}
\alias{blms_model}
\title{blms_model}
\usage{
blms_model(
  formula,
  ...,
  data = NULL,
  formula_only = !compile,
  compile = run,
  run = F,
  model_class = NULL,
  model_spec = NULL,
  par_form = NULL,
  par_transform = NULL
)
}
\arguments{
\item{formula}{response formula specifying the response and its prediction}

\item{...}{You can pass further formulas to define predictions of parameters
used in the response formula. Any non-formula objects passed via
\code{...} are passed on to other functions (i.e., \code{brmsformula} and
\code{brm})}

\item{data}{the data containing the response variable and any variables used
for prediction. Any variables that occur in formulas but cannot be
found in \code{data} will be considered non-linear parameters.}

\item{formula_only}{return the \code{brmsformula} created from inputs but do not
(try to) compile or run the model}

\item{compile}{compile the model and return a \code{brmsfit} object. If \code{run} is
\code{FALSE} \code{brm} is called with its \code{chains} argument set to \code{0} which
will compile the model but will not sample}

\item{run}{calls \code{brm} with the \code{brmsformula} created from inputs, \code{data},
\code{stanvars} defined by \code{model_spec} or \code{stanvars} defined in \code{dots},
and any other arguments passed vie \code{...} that were not consumed for
creating the formula}

\item{model_class}{character of length 1 specifying the class of the model
(see \code{names(get_all_model_specs())} for names of implemented model
classes)}

\item{model_spec}{list defining a model specification (see
\code{get_all_model_specs()} for currently implemented models and their
specifications; you can supply other model specifications that
follow the same structure)}

\item{par_form}{\code{list()} object containing \code{formula} object(s) that specify
prediction formulas for other parameters; formulas passed via \code{...}
will override formulas passed via \code{par_form}}

\item{par_transform}{\code{list()} object containing transformation formulas for
other formulas of the model; transformation formulas can e.g. be used
to transform parameters that have upper and/or lower bounds. When
passing transformation formulas via \code{par_transform} you can specify
the parameter to transform on the left-hand side and the
transformation formula - containing the same parameter - on the
right-hand-side of the formula (e.g. \code{d ~ inv_logit(d)}). You can also
pass transformation formulas via the \code{...} by using a
transformation statement (\code{transform()}) on the left-hand side of the
formula and the transformation statement on the right-hand side of
the formula, e.g. \code{transform(d) ~ inv_logit(d)}. \code{blms} creates from
this formula a non-linear formula in which the parameter on the
right-hand side is replaced by a new parameter named after the
original parameter with \code{'raw'} as suffix (e.g.
\code{d ~inv_logit(draw)}). If in \code{par_form} there is a formula
that has the respective parameter (\code{d} in the current example) on its
left-hand side, then in that formula the parameter on the left-hand
side is renamed by adding the suffix \code{'raw'} to the first term.
If you want to use
the same transformation for several parameters you can specify
multiple parameters on the left-hand side combining them by \code{+} and
use 'x' on the right-hand side of the formula. This will create as
many transformation formulas as there are variables on the left-hand
side and replace 'x' on the right-hand side.}
}
\value{
If \code{formula_only} is \code{TRUE} then the return value is an object of
class \code{blmsformula} (inheriting from class \code{brmsformula}), else the
return value is of class \code{blmsfit} (inheriting from class \code{brmsfit})
}
\description{
blms_model
}
