% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsformula.R
\name{blmsformula}
\alias{blmsformula}
\title{Create a blmsformula object}
\usage{
blmsformula(
  formula,
  ...,
  data = NULL,
  model_class = NULL,
  model_spec = NULL,
  par_form = NULL,
  par_transform = NULL,
  model_func_has_blockgrp = F
)
}
\arguments{
\item{formula}{a \code{formula} defining response and prector variables (and
optionally a block term)}

\item{...}{other formula objects describing prediction of model parameters}

\item{data}{optional, if given, a \code{data.frame} or the like that contains
the data for the model to be fitted with the formula}

\item{model_class}{character defining the model class}

\item{model_spec}{\code{model_spec} as returned by \code{get_model_spec(model_class)}}

\item{par_form}{(linear) formulas for the prediction of model parameters
defined by the model specified by \code{model_class} and/or \code{model_spec}}

\item{par_transform}{\code{formula} objects defining the transformation aplied to
the linear formula of the parameter named on the left-hand side of
the formula. Other formulas for the respective parameter will be
changed by appending \code{var} to the parameter name and the}

\item{model_func_has_blockgrp}{}
}
\value{
object of class \code{blmsformula} which inherits from \code{brmsformula}
}
\description{
Create a blmsformula object
}
