% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-helper.R
\name{create_doc_formula_arg}
\alias{create_doc_formula_arg}
\title{Create default description for \code{formula} argument of a
\code{blms_model}}
\usage{
create_doc_formula_arg(
  model_class = "",
  response_desc = "",
  pred_desc = "",
  block_restart = ""
)
}
\arguments{
\item{model_class}{\code{character}. Class name.}

\item{response_desc}{\code{character}. Description of how the response
variable needs to be coded.}

\item{pred_desc}{\code{character}. Description of how the prediction
variable(s) need to be coded.}

\item{block_restart}{\code{character}. Description of what happens when a
block is restarted}
}
\value{
A (formatted) \code{character} string containing the description
of the \code{formula} argument for e \code{blms_model}.
}
\description{
Create default description for \code{formula} argument of a
\code{blms_model}
}
\examples{
create_doc_formula_arg()
}
