% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsformula.R
\name{default_prior.blmsformula}
\alias{default_prior.blmsformula}
\title{default_prior implementation for blmsformula}
\usage{
\method{default_prior}{blmsformula}(
  object,
  data,
  family = gaussian(),
  autocor = NULL,
  data2 = NULL,
  knots = NULL,
  drop_unused_levels = TRUE,
  sparse = NULL,
  ...
)
}
\description{
default_prior implementation for blmsformula
}
