% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_model_par.R
\name{df_copy_classes}
\alias{df_copy_classes}
\title{Try to copy classes of variables from another data.frame}
\usage{
df_copy_classes(df, df2, unlist = T)
}
\arguments{
\item{df}{\code{data.frame} to copy classes of variables to}

\item{df2}{\code{data.frame} to take variable classes from}
}
\value{
\code{df} with variables found in \code{df2} coerced to their class
}
\description{
Try to copy classes of variables from another data.frame
}
