% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsformula-helpers.R
\name{find_bar_terms}
\alias{find_bar_terms}
\title{Find and return terms of a formula that (do not) contain a bar (|)}
\usage{
find_bar_terms(form, invert = F)
}
\arguments{
\item{form}{A \code{formula} object}

\item{invert}{Return term that do not contain a \code{|}}
}
\value{
A \code{character} vector of the selected terms
}
\description{
Find and return terms of a formula that (do not) contain a bar (|)
}
