% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsformula-helpers.R
\name{formula_replace}
\alias{formula_replace}
\title{Replace patterns in part of a formula}
\usage{
formula_replace(form, side, pattern, replacement, perl = T, ...)
}
\arguments{
\item{form}{A \code{formula} object.}

\item{side}{\code{'left'}, \code{'right'}, or \code{'both'} for the
left-hand side, the right-hand side, or both sides of a formula,
respectively. Can be abbreviated.}

\item{pattern}{character string containing a \link[base]{regular expression}
    (or character string for \code{fixed = TRUE}) to be matched
    in the given character vector.  Coerced by
    \code{\link[base]{as.character}} to a character string if possible.  If a
    character vector of length 2 or more is supplied, the first element
    is used with a warning.  Missing values are allowed except for
    \code{regexpr}, \code{gregexpr} and \code{regexec}.}

\item{replacement}{a replacement for matched pattern in \code{sub} and
    \code{gsub}.  Coerced to character if possible.  For
    \code{fixed = FALSE} this can include backreferences \code{"\1"} to
    \code{"\9"} to parenthesized subexpressions of \code{pattern}.  For
    \code{perl = TRUE} only, it can also contain \code{"\U"} or
    \code{"\L"} to convert the rest of the replacement to upper or
    lower case and \code{"\E"} to end case conversion.  If a
    character vector of length 2 or more is supplied, the first element
    is used with a warning.  If \code{NA}, all elements in the result
    corresponding to matches will be set to \code{NA}.
  }

\item{perl}{logical.  Should Perl-compatible regexps be used?}
}
\value{
A \code{formula} object with \code{pattern} replaced by
\code{replacement} on the side of the \code{formula} defined by
\code{side}.
}
\description{
Replace patterns in part of a formula
}
