% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-specifications.R
\name{get_all_model_specs}
\alias{get_all_model_specs}
\title{Get currently available model specifications}
\usage{
get_all_model_specs()
}
\value{
a \code{list()} of model specifications for blms models
}
\description{
Get currently available model specifications
}
