% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_model_par.R
\name{get_all_re_vars}
\alias{get_all_re_vars}
\title{Get all variables from group-level effects of all formulas in brmsformula}
\usage{
get_all_re_vars(formula, nlpar = NULL, group = T, effect = T)
}
\arguments{
\item{formula}{(non-linear) \code{brmsformula}}

\item{nlpar}{list of non-linear parameters to obtain re vars for}

\item{group}{return grouping variables for selected \code{nlpar}s}

\item{effect}{return effect variables for selected \code{nlpar}s}
}
\value{
character vector of variable names used in group-level terms
}
\description{
Get all variables from group-level effects of all formulas in brmsformula
}
