% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-helper.R
\name{get_doc_description_md}
\alias{get_doc_description_md}
\title{Get description for function documentation}
\usage{
get_doc_description_md(ms)
}
\arguments{
\item{ms}{A \code{model_spec} object.}
}
\value{
Description string formatted for roxygen2 md
}
\description{
Get description for function documentation
}
\examples{
get_doc_description_md(model_spec)
}
