% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-helper.R
\name{get_doc_details_md}
\alias{get_doc_details_md}
\title{Details string formatted for roxygen2 md}
\usage{
get_doc_details_md(ms)
}
\arguments{
\item{ms}{A \code{model_spec} object.}
}
\value{
Details string formatted for roxygen2 md
}
\description{
Details string formatted for roxygen2 md
}
\examples{
get_doc_details_md(model_spec)
}
