% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-helper.R
\name{get_doc_formula_arg_md}
\alias{get_doc_formula_arg_md}
\title{Get \code{formula} argument description for function documentation}
\usage{
get_doc_formula_arg_md(ms)
}
\arguments{
\item{ms}{A \code{model_spec} object.}
}
\value{
Description string for \code{formula} argument
formatted for roxygen2 md
}
\description{
Get \code{formula} argument description for function documentation
}
\examples{
get_doc_formula_arg_md(model_spec)
}
