% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-helper.R
\name{get_doc_title_md}
\alias{get_doc_title_md}
\title{Get title for function documentation}
\usage{
get_doc_title_md(ms)
}
\arguments{
\item{ms}{A \code{model_spec} object.}
}
\value{
Title string formatted for roxygen2 md
}
\description{
Get title for function documentation
}
\examples{
get_doc_title_md(model_spec)
}
