% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsformula-helpers.R
\name{get_formula}
\alias{get_formula}
\title{Get the \code{formula} from. \code{x}}
\usage{
get_formula(x)
}
\arguments{
\item{x}{}
}
\value{
a \code{formula} object if \code{is_formula(x) == TRUE} else
\code{NULL}
}
\description{
Get the \code{formula} from. \code{x}
}
