% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-specifications.R
\name{get_model_spec}
\alias{get_model_spec}
\title{Get model specification for a specific model class}
\usage{
get_model_spec(class)
}
\arguments{
\item{class}{character giving the name of the model class}
}
\value{
a model specification for a blms model
}
\description{
Get model specification for a specific model class
}
