% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_model_par.R
\name{get_nlpar_draws}
\alias{get_nlpar_draws}
\title{Get predicted draws for non-linear parameters}
\usage{
get_nlpar_draws(
  fit,
  nlpar,
  by = ~1,
  predictor_values = list(),
  re_formula = NA,
  method = c("epred", "linpred", "predict"),
  new_data = NA
)
}
\arguments{
\item{fit}{A \code{brmsfit}object or inheritor thereof}

\item{nlpar}{The non-linear parameter for which draws should be obtained}

\item{by}{Either one-sided formula specifying predictors of the non-linear
parameter that should be included, or a character vector of names of
those predictors. For factors, draws will be obtained separately for
each level, for continuous variables, by default, draws will be
obtained for mean-sd, mean, and mean+sd (can be adjusted/replaced by
defining specific values for the predictors in \code{predictor_values}).
For variables not named here (or in \code{predictor_values}), if they
are factors, draws will be averaged across all levels, for
continuous variables only the mean will be used to obtain draws.}

\item{predictor_values}{a named \code{list()} with names corresponding to
predictor variables of the non-linear parameter, and their values
values of the predictor variable for which draws should be obtained}

\item{re_formula}{formula for group-level effects to be included}

\item{method}{\code{'epred'} and \code{'linpred'} should be the same for non-linear
parameters and account for uncertainty of the fixed coefficients and
uncertainty of the variance parameters of the groups. \code{'predict'}
additionally accounts for uncertainty for each individual observation
(e.g., observational-level residual variance)}
}
\value{
\code{data.frame} with columns for the predictor variables containing
the values used to obtain draws and a column named after the
non-linear parameter respective
}
\description{
Get predicted draws for non-linear parameters
}
