% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_model_par.R
\name{get_nlpar_pred_grid}
\alias{get_nlpar_pred_grid}
\title{Get prediction grid for posterior draws of a non-linear parameter}
\usage{
get_nlpar_pred_grid(
  fit,
  nlpar,
  by = ~1,
  re_formula = NA,
  predictor_values = list()
)
}
\arguments{
\item{fit}{A \code{brmsfit}object or inheritor thereof}

\item{nlpar}{The non-linear parameter for which draws should be obtained}

\item{by}{Either one-sided formula specifying predictors of the non-linear
parameter that should be included, or a character vector of names of
those predictors. For factors, draws will be obtained separately for
each level, for continuous variables, by default, draws will be
obtained for mean-sd, mean, and mean+sd (can be adjusted/replaced by
defining specific values for the predictors in \code{predictor_values}).
For variables not named here (or in \code{predictor_values}), if they
are factors, draws will be averaged across all levels, for
continuous variables only the mean will be used to obtain draws.}

\item{re_formula}{formula for group-level effects to be included}

\item{predictor_values}{a named \code{list()} with names corresponding to
predictor variables of the non-linear parameter, and their values
values of the predictor variable for which draws should be obtained}
}
\value{
A \code{data.frame} expanding combinations of variables as defined by
the inputs. The main purpose is to use this to pbtain some kind of
expected or predicted draws.
}
\description{
Get prediction grid for posterior draws of a non-linear parameter
}
