% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_model_par.R
\name{get_nlpar_predictors}
\alias{get_nlpar_predictors}
\title{Get All Predictors for a Nonlinear Parameter}
\usage{
get_nlpar_predictors(nlpar_name, formula_obj, data_vars, seen = character())
}
\arguments{
\item{nlpar_name}{Name of the nonlinear parameter.}

\item{formula_obj}{A brms formula object.}

\item{data_vars}{Character vector of variable names in the data.}

\item{seen}{Internal recursion tracker.}
}
\value{
A character vector of required variable names.
}
\description{
Recursively identifies all variables (excluding group-level terms) needed to compute a specific nonlinear parameter.
}
\keyword{internal}
