% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-specifications.R
\name{get_parameter_names}
\alias{get_parameter_names}
\title{Get the names of the parameters fitted by a blms model of a specific class}
\usage{
get_parameter_names(model_spec)
}
\arguments{
\item{model_spec}{model specification as returned by \code{get_model_spec()} (see
\code{get_all_model_specs()} for a list of currently defined models)}
}
\value{
character vector containing the names of the parameters fitted by
a blms model of a specific class
}
\description{
Get the names of the parameters fitted by a blms model of a specific class
}
