% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_model_par.R
\name{get_used_nlpars}
\alias{get_used_nlpars}
\title{get nlpars used in other nlpar formula}
\usage{
get_used_nlpars(formula, nlpar)
}
\arguments{
\item{formula}{\code{brmsformula}}

\item{nlpar}{\code{nlpar} for which used nlpars should be return}
}
\value{
character vector
}
\description{
get nlpars used in other nlpar formula
}
