% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsformula-helpers.R
\name{is.formula}
\alias{is.formula}
\title{Test if input is a formula}
\usage{
is.formula(x)
}
\arguments{
\item{x}{Any object}
}
\value{
\code{TRUE} if \code{x} is a formula, else \code{FALSE}
}
\description{
Test if input is a formula
}
