% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsformula-helpers.R
\name{is.twosided}
\alias{is.twosided}
\title{Test whether a formula is two-sided}
\usage{
is.twosided(form)
}
\arguments{
\item{form}{The \code{formula}}
}
\value{
\code{TRUE} if the \code{formula} is two-sided, that is,
if it has a left-hand side
}
\description{
Test whether a formula is two-sided
}
