% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsformula-helpers.R
\name{is_formula}
\alias{is_formula}
\title{Test if input is \code{formula} or \code{list}
(i.e. \code{brms::nlf, brms::lf}) defining one}
\usage{
is_formula(x)
}
\arguments{
\item{x}{An to be tested for its \code{formula}}
}
\value{
\code{TRUE} if \code{x} is a formula (also returns \code{TRUE} for
\code{brms::nlf} and \code{brms::lf} objects), else \code{FALSE}
}
\description{
Test if input is \code{formula} or \code{list}
(i.e. \code{brms::nlf, brms::lf}) defining one
}
