% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsformula-helpers.R
\name{lhs}
\alias{lhs}
\title{Get the left-hand side of a formula}
\usage{
lhs(f)
}
\arguments{
\item{f}{A \code{formula} object}
}
\value{
The left-hand side of the formula, if it has one,
otherwise \code{NULL}
}
\description{
Get the left-hand side of a formula
}
