% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsformula-helpers.R
\name{obj_set_env}
\alias{obj_set_env}
\title{Set specific environment for an object}
\usage{
obj_set_env(obj, envir)
}
\arguments{
\item{obj}{An object.}

\item{envir}{An \code{environment} object.}
}
\value{
\code{obj} with attribute \code{'.Environment'} set to \code{envir}
}
\description{
Set specific environment for an object
}
