% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-helper.R
\name{parameters_md_list}
\alias{parameters_md_list}
\title{Get list of model parameters for details section of roxygen2 md}
\usage{
parameters_md_list(ms)
}
\arguments{
\item{ms}{A \code{model_spec} object.}
}
\value{
Formatted list for detals section of roxygen2 md
}
\description{
Get list of model parameters for details section of roxygen2 md
}
\examples{
parameters_md_list(model_spec)
}
