% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsformula.R
\name{parse_blms_model_inputs}
\alias{parse_blms_model_inputs}
\title{Parse inputs to \code{brmsformula} or \code{blms_model}}
\usage{
parse_blms_model_inputs(
  formula,
  ...,
  data = NULL,
  model_class = NULL,
  model_spec = NULL,
  par_form = NULL,
  par_transform = NULL,
  model_func_has_blockgrp = F
)
}
\arguments{
\item{formula}{A \code{formula} defining response and predictor variables (and
optionally a block term)}

\item{...}{Other formula objects describing prediction of model parameters}

\item{data}{Optional, if given, a \code{data.frame} or the like that contains
the data for the model to be fitted with the formula}

\item{model_class}{A character defining the model class}

\item{model_spec}{\code{model_spec} as returned by \code{get_model_spec(model_class)}}

\item{par_form}{(linear) formulas for the prediction of model parameters
defined by the model specified by \code{model_class} and/or \code{model_spec}}

\item{par_transform}{\code{formula} objects defining the transformation applied to
the linear formula of the parameter named on the left-hand side of
the formula. Other formulas for the respective parameter will be
changed by appending \code{var} to the parameter name and the}

\item{model_func_has_blockgrp}{Currently not used.}
}
\value{
\code{list()} with slots \code{model_class}, \code{model_spec},
\code{response_formula}, \code{parameter_formulas},
\code{resp_var}, \code{pred_vars}, \code{block_vars},
\code{trial_var}, \code{cond_var}, \code{dec_var},
\code{brmsformula},
}
\description{
Parse inputs to \code{brmsformula} or \code{blms_model}
}
