% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsformula-helpers.R
\name{parse_par_formulas}
\alias{parse_par_formulas}
\title{Parse parameter and transformation formulas}
\usage{
parse_par_formulas(par_forms, par_transforms, model_pars = NULL, data = NULL)
}
\arguments{
\item{par_forms}{A \code{list()} of \code{formula} objects}

\item{par_transforms}{A \code{list()} of \code{formula} objects}

\item{model_pars}{A character vector defining the names of model parameters.
Currently not used.}

\item{data}{A \code{data.frame} (or the like) used to check whether all the
variables on the right-hand side of a formula in \code{par_forms} is
present in the data. If not, the formula is returned as a non-linear
formula}
}
\value{
A \code{list()} of linear and non-linear formulas to be passed to
\code{brms::brmsformula}
}
\description{
Parse parameter and transformation formulas
}
