% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-helper.R
\name{parse_param_name}
\alias{parse_param_name}
\title{Parse parameter names of a model to Latex code}
\usage{
parse_param_name(x)
}
\arguments{
\item{x}{A \code{character}.}
}
\value{
\code{x} with greek letters and subscripts formatted for
\code{"\\eqn{}"}
}
\description{
Parse parameter names of a model to Latex code
}
\examples{
parse_param_name('alphapos')
}
