% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_par.R
\name{plot_model_par}
\alias{plot_model_par}
\title{Plot Posterior Predictions for a Nonlinear Parameter in a \code{brms} Model}
\usage{
plot_model_par(
  fit,
  nlpar,
  by,
  method = c("epred", "linpred", "predict"),
  re_formula = NA,
  predictor_values = list(),
  width = c(0.5, 0.8, 0.95)
)
}
\arguments{
\item{fit}{A \code{brmsfit} model object.}

\item{nlpar}{A string indicating the name of the nonlinear parameter (e.g., \code{"gamma"}).}

\item{method}{One of \code{"epred"} or \code{"linpred"} to specify which type of prediction to plot.}

\item{re_formula}{Re formula passed to \code{posterior_epred()} / \code{posterior_linpred()}. Use \code{NA} to exclude random effects.}

\item{predictor_values}{A named list providing custom values for predictors in the prediction grid.}

\item{width}{Numeric vector of credible intervals passed to \code{ggdist::stat_halfeye()}.}

\item{formula}{A formula specifying which predictor(s) to vary in the plot (e.g., \code{~ condition}).}
}
\value{
A \code{ggplot} object showing the posterior distributions of the specified nonlinear parameter.
}
\description{
This function visualizes the posterior predictions (or linear predictions) of a specified
non-linear parameter (\code{nlpar}) in a \code{brmsfit} model. It infers variables that affect the
\code{nlpar}, constructs a prediction grid, and plots posterior distributions using \code{ggdist}.
}
