% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blms_model.R
\name{print.blmsfit}
\alias{print.blmsfit}
\title{Print method for class blmsfit}
\usage{
\method{print}{blmsfit}(x, digits = 2, ...)
}
\arguments{
\item{...}{}

\item{fit}{An object of class \code{blmsfit}}
}
\description{
Print method for class blmsfit
}
