% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-specifications.R
\name{print.model_spec}
\alias{print.model_spec}
\title{print method for \code{model_spec}}
\usage{
\method{print}{model_spec}(ms)
}
\arguments{
\item{ms}{A \code{model_spec} object.}
}
\description{
print method for \code{model_spec}
}
