% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsformula-helpers.R
\name{rhs}
\alias{rhs}
\title{Get the right hand side of a formula}
\usage{
rhs(f)
}
\arguments{
\item{f}{}
}
\value{
The right-hand side of a formula
}
\description{
Get the right hand side of a formula
}
