% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsformula-helpers.R
\name{rhs_dp}
\alias{rhs_dp}
\title{Get the deparsed version of left-hand side of a formula}
\usage{
rhs_dp(f, replace_ws = T, trim_ws = T, remove_ws = F)
}
\arguments{
\item{f}{A \code{formula} object}

\item{replace_ws}{\code{logical}. Replaces all kinds of extended whitespaces
(e.g. \code{\t}, \code{\n}, or multiple whitespaces) by a single space}

\item{trim_ws}{removes whitespaces from the beginniing and the end of a
string}

\item{remove_ws}{removes any whitespaces from the string}
}
\value{
The right-hand side of a formula, possibly with whitespaces
replaced or removed
}
\description{
Get the deparsed version of left-hand side of a formula
}
