% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsformula-helpers.R
\name{rm_ws}
\alias{rm_ws}
\title{Remove whitespaces from a string}
\usage{
rm_ws(str, replace_ws = T, trim_ws = T, remove_ws = F)
}
\arguments{
\item{str}{A \code{character()}}

\item{replace_ws}{\code{logical}. Replaces all kinds of extended whitespaces
(e.g. \code{\t}, \code{\n}, or multiple whitespaces) by a single space}

\item{trim_ws}{removes whitespaces from the beginniing and the end of a
string}

\item{remove_ws}{removes any whitespaces from the string}
}
\value{
The input \code{character} with whitespaces replaced or removed
}
\description{
Remove whitespaces from a string
}
