% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsformula-helpers.R
\name{set_formula}
\alias{set_formula}
\title{Set the \code{formula} of \code{x}}
\usage{
set_formula(x, new_x)
}
\arguments{
\item{x}{\code{formula}}

\item{new_x}{new \code{formula}}
}
\value{
\code{new_x} with attributes of \code{x}
}
\description{
Set the \code{formula} of \code{x}
}
