% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsformula-helpers.R
\name{split_flist_by_lhs}
\alias{split_flist_by_lhs}
\title{Split each formula in a list into multiple formulas based on variables
defined in the left-hand side of the formula}
\usage{
split_flist_by_lhs(flist)
}
\arguments{
\item{flist}{A \code{list()} containing formulas, each possibly containing
multiple variables (typically combined by \code{+}) on its
left-hand side}
}
\value{
A \code{list()} containing \code{formula} objects created by
splitting the input formulas into multiple formulas if multiple
variables are listed on their left-hand side
}
\description{
Split each formula in a list into multiple formulas based on variables
defined in the left-hand side of the formula
}
\details{
Variables on the left-hand side of the formula are extracted using
\code{all.vars}
}
\examples{
split_flist_by_lhs(list(a + b ~ c, d ~ e))

}
