% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsformula.R
\name{stancode.blmsformula}
\alias{stancode.blmsformula}
\title{stancode implementation for blmsformula}
\usage{
\method{stancode}{blmsformula}(
  object,
  data,
  family = gaussian(),
  prior = NULL,
  autocor = NULL,
  data2 = NULL,
  cov_ranef = NULL,
  sparse = NULL,
  sample_prior = "no",
  stanvars = NULL,
  stan_funs = NULL,
  knots = NULL,
  drop_unused_levels = TRUE,
  threads = getOption("brms.threads", NULL),
  normalize = getOption("brms.normalize", TRUE),
  save_model = NULL,
  ...
)
}
\description{
stancode implementation for blmsformula
}
