% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blms_model.R
\name{standata.blmsfit}
\alias{standata.blmsfit}
\title{standata implementation for blmsfit}
\usage{
\method{standata}{blmsfit}(
  object,
  newdata = NULL,
  re_formula = NULL,
  newdata2 = NULL,
  new_objects = NULL,
  incl_autocor = TRUE,
  ...
)
}
\description{
standata implementation for blmsfit
}
