% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsformula.R
\name{standata.blmsformula}
\alias{standata.blmsformula}
\title{standata implementation for blmsformula}
\usage{
\method{standata}{blmsformula}(
  object,
  data,
  family = gaussian(),
  prior = NULL,
  autocor = NULL,
  data2 = NULL,
  cov_ranef = NULL,
  sample_prior = "no",
  stanvars = NULL,
  threads = getOption("brms.threads", NULL),
  knots = NULL,
  drop_unused_levels = TRUE,
  ...
)
}
\description{
standata implementation for blmsformula
}
