% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_model_par.R
\name{summarize_model_par}
\alias{summarize_model_par}
\title{Get summary of a non-linear parameter}
\usage{
summarize_model_par(fit, nlpar, ...)
}
\arguments{
\item{fit}{An object of class \code{brmsfit}}

\item{nlpar}{The name of the non-linear parameter}

\item{...}{}
}
\value{
A \code{data.frame} with summaries of the non-linear parameter
}
\description{
Get summary of a non-linear parameter
}
