% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-helper.R
\name{tabular}
\alias{tabular}
\title{Create md formatted table from \code{data.frame}}
\usage{
tabular(df, ...)
}
\arguments{
\item{df}{A \code{data.frame}}

\item{...}{Formatting options.}
}
\value{
Formatted text representation of the input \code{data.frame}.
}
\description{
Create md formatted table from \code{data.frame}
}
\details{
Copied from roxygen2 documentation
}
\examples{
cat(tabular(mtcars[1:5, 1:5]))
}
