% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsformula.R
\name{validate_data}
\alias{validate_data}
\title{Validate data for fitting with blmsformula}
\usage{
validate_data(data, blmsformula)
}
\arguments{
\item{data}{A \code{data.frame}containing the data to be validated.}

\item{blmsformula}{A \code{blmsformula} object.}
}
\value{
A validated \code{data.frame}based on \code{blmsformula}
}
\description{
Validate data for fitting with blmsformula
}
