% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmsformula-helpers.R
\name{validate_var}
\alias{validate_var}
\title{Validate if a variable is suitable to be used for parameter in a model}
\usage{
validate_var(var, par, data = NULL, valid_values = NULL, null_ok = F)
}
\arguments{
\item{var}{The variable. If \code{length(variable)>1} only the first element
is used.}

\item{par}{Name of the parameter for which the variable is validated
(only used for error messages)}

\item{data}{Optional. A \code{data.frame} (or the like) in which to look for a
variable named \code{var}}

\item{valid_values}{Only applies if \code{data} is not \code{NULL} and a
variable named \code{var} was found in \code{data}.
If \code{valid_values} is a vector (i.e.
\code{is.vector(valies_values)} is \code{TRUE}), then values of
\code{var} are checked against values in \code{valid_values} using
\code{setdiff}. \code{valid_values} can also be a function that takes
values of \code{var} as input. If all values are deemed to be valid, it
should return those values. Otherwise, the function can raise an
error or return an error message as a \code{character} of length 1
with a class attribute of 'error_message' (if that is not the case
values are considered to be valid).}

\item{null_ok}{If \code{is.null(var)} return \code{NULL} if
\code{null_ok = TRUE}, else (the default) raise an error}
}
\value{
\code{var}, unless an error is raised.
}
\description{
Validate if a variable is suitable to be used for parameter in a model
}
